Fix pagination bug
The bug is when user reach the end of the list, the app force closed.

This is because the PagingSource set the same nextKey twice which is not allowed.

The preventation is by checking the current key is equals to last key. If it's the set the nextKey to null.